requirements.txt
numpy==1.14.3
phue==1.0
pillow==5.0.0
python-dotenv==0.8.2
scikit-learn==0.19.1
scipy==1.0.1
